A modded Waila acts the way Mobius and I discussed
This jar contains the 1.5.2 dev branch from Waila, with minor modifications
related to ModLoader as recommended by Professor Mobius. It also contains
a change to the order in which Waila runs the basic handlers in relation to
custom-registered handlers. The builtin handlers now run _before_ mod-registered
handlers, allowing the mod-registered handlers to override them. This is
something that hopefully will be possible in Waila in the future to produce
the effects contained in this branch.